% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_scatter_plot.R
\name{plot_scatter_box}
\alias{plot_scatter_box}
\title{Get scatter plot from input dataframe}
\usage{
plot_scatter_box(dt_sb, var_x, var_y, col_g, xlab, ylab, title_lab)
}
\description{
Get scatter plot from input dataframe
}
