% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{generate_dna_report}
\alias{generate_dna_report}
\title{Generate Quartet Genomics report}
\usage{
generate_dna_report(
  qc_result,
  report_template,
  report_dir = NULL,
  report_name = NULL
)
}
\arguments{
\item{qc_result}{list}

\item{report_template}{character}

\item{report_dir}{character}

\item{report_name}{character}
}
\description{
Use calculated DNA result to generate QC report
}
