% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_report.R
\name{generate_dna_report}
\alias{generate_dna_report}
\title{Generate Quartet Genomics report}
\usage{
generate_dna_report(
  DNA_result = NULL,
  doc_file_path = NULL,
  output_path = NULL
)
}
\arguments{
\item{DNA_result}{list}

\item{output_path}{character}

\item{temp_doc_path}{character}
}
\value{
word file
}
\description{
Use calculated Met result to generate report
}
